refactor: XSSFRow 매핑 리팩터링
