+++ growthmate-batch/src/main/java/com/growup/growthmate/support/XSSFRowMapper.java
@@ -1,4 +1,4 @@
-package com.growup.growthmate;
+package com.growup.growthmate.support;
 
 import org.apache.poi.xssf.usermodel.XSSFRow;
